Version 2.1.5
git-svn-id: https://plugins.svn.wordpress.org/custom-login/trunk@715758 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ classes/class.settings-api.php
@@ -14,7 +14,7 @@ if ( !class_exists( 'Extendd_Plugin_Settings_API' ) ):
 	/**
 	 * Version
 	 */
-	var $api_version = '1.0.6';
+	var $api_version = '1.0.7';
 
     /**
      * settings sections array
@@ -775,7 +775,7 @@ if ( !class_exists( 'Extendd_Plugin_Settings_API' ) ):
 			}
 		}
 			
-		if ( trim( $old_message ) !== trim( $announcement->message ) ) {
+		if ( trim( $old_message ) !== trim( $announcement->message ) && !empty( $old_message ) ) {
 			delete_user_meta( get_current_user_id(), $ignore, 1 );
 			delete_transient( $transient );
 			delete_option( $this->prefix . '_announcement_message' );
